+++ OpenSim/Server/Handlers/Inventory/XInventoryInConnector.cs
@@ -600,7 +600,7 @@ namespace OpenSim.Server.Handlers.Asset
 
             folder.ParentID =  new UUID(data["ParentID"].ToString());
             folder.Type = short.Parse(data["Type"].ToString());
-            folder.Version = ushort.Parse(data["Version"].ToString());
+            folder.Version = uint.Parse(data["Version"].ToString());
             folder.Name = data["Name"].ToString();
             folder.Owner =  new UUID(data["Owner"].ToString());
             folder.ID = new UUID(data["ID"].ToString());
