+++ OpenSim/Services/Connectors/Inventory/XInventoryServicesConnector.cs
@@ -500,7 +500,7 @@ namespace OpenSim.Services.Connectors
             {
                 folder.ParentID = new UUID(data["ParentID"].ToString());
                 folder.Type = short.Parse(data["Type"].ToString());
-                folder.Version = ushort.Parse(data["Version"].ToString());
+                folder.Version = uint.Parse(data["Version"].ToString());
                 folder.Name = data["Name"].ToString();
                 folder.Owner = new UUID(data["Owner"].ToString());
                 folder.ID = new UUID(data["ID"].ToString());
